Rename zul/wgt to TS
