+++ zul/src/main/resources/web/js/zul/wgt/Area.ts
@@ -1,4 +1,4 @@
-/* Area.js
+/* Area.ts
 
 	Purpose:
