+++ zul/src/main/resources/web/js/zul/wgt/ButtonRenderer.ts
@@ -1,4 +1,4 @@
-/* ButtonRenderer.js
+/* ButtonRenderer.ts
 
 	Purpose:
