+++ zul/src/main/resources/web/js/zul/wgt/Captcha.ts
@@ -1,4 +1,4 @@
-/* Captcha.js
+/* Captcha.ts
 
 	Purpose:
