+++ zul/src/main/resources/web/js/zul/wgt/Chart.ts
@@ -1,4 +1,4 @@
-/* Chart.js
+/* Chart.ts
 
 	Purpose:
