+++ zul/src/main/resources/web/js/zul/wgt/Fileupload.ts
@@ -1,4 +1,4 @@
-/* Fileupload.js
+/* Fileupload.ts
 
 	Purpose:
