+++ zul/src/main/resources/web/js/zul/wgt/Html.ts
@@ -1,4 +1,4 @@
-/* Html.js
+/* Html.ts
 
 	Purpose:
