--- zul/src/main/resources/web/js/zul/wgt/Idspace.js
@@ -1,26 +1,0 @@
-/* Idspace.js
-
-	Purpose:
-
-	Description:
-
-	History:
-		Thu Nov 03 12:15:49     2011, Created by benbai
-
-Copyright (C) 2011 Potix Corporation. All Rights Reserved.
-
-This program is distributed under LGPL Version 3.0 in the hope that
-it will be useful, but WITHOUT ANY WARRANTY.
-*/
-/**
- * Just like DIV tag but implements IdSpace.
- * @since 6.0.0
- */
-zul.wgt.Idspace = zk.$extends(zul.wgt.Div, {
-	$init: function () {
-		this._fellows = {};
-		this.$supers('$init', arguments);
-	}
-});
-//ZK-3255: support nodom mold
-zk.$intercepts(zul.wgt.Idspace, zk.NoDOM);
