+++ zul/src/main/resources/web/js/zul/wgt/Imagemap.ts
@@ -1,4 +1,4 @@
-/* Imagemap.js
+/* Imagemap.ts
 
 	Purpose:
