+++ zul/src/main/resources/web/js/zul/wgt/Include.ts
@@ -1,4 +1,4 @@
-/* Include.js
+/* Include.ts
 
 	Purpose:
