+++ zul/src/main/resources/web/js/zul/wgt/Rating.ts
@@ -1,4 +1,4 @@
-/* Rating.js
+/* Rating.ts
 
 	Purpose:
