+++ zul/src/main/resources/web/js/zul/wgt/index.ts
@@ -1,10 +1,10 @@
-// export * from './ButtonRenderer'; // jsdoc="true"
+export * from './ButtonRenderer'; // jsdoc="true"
 export * from './A';
 export * from './Cell';
 export * from './Div';
 export * from './Span';
-// export * from './Idspace';
-// export * from './Include';
+export * from './Idspace';
+export * from './Include';
 export * from './Label';
 export * from './Button';
 export * from './Separator';
@@ -12,22 +12,22 @@ export * from './Space';
 export * from './Caption';
 export * from './Checkbox';
 export * from './Groupbox';
-// export * from './Html';
+export * from './Html';
 export * from './Popup';
 export * from './Radio';
 export * from './Radiogroup';
 export * from './Toolbar';
 export * from './Toolbarbutton';
-// export * from './Nodom';
+export * from './Nodom';
 export * from './Image';
-// export * from './Imagemap';
-// export * from './Area';
-// export * from './Chart';
-// export * from './Captcha';
+export * from './Imagemap';
+export * from './Area';
+export * from './Chart';
+export * from './Captcha';
 export * from './Progressmeter';
-// export * from './Fileupload';
-// export * from './Combobutton';
-// export * from './Selectbox';
+export * from './Fileupload';
+export * from './Combobutton';
+export * from './Selectbox';
 export * from './Notification';
-// export * from './Rating';
+export * from './Rating';
 export * from './Inputgroup';
