+++ zul/src/main/resources/web/js/zul/wgt/zk.wpd
@@ -12,37 +12,38 @@ Copyright (C) 2008 Potix Corporation. All Rights Reserved.
 -->
 
 <package name="zul.wgt" language="xul/html" depends="zul">
-	<script src="ButtonRenderer.js" jsdoc="true"/>
-	<widget name="A"/>
-	<widget name="Cell"/>	
-	<widget name="Div"/>
-	<widget name="Span"/>
-	<widget name="Idspace"/>
-	<widget name="Include"/>
-	<widget name="Label"/>
-	<widget name="Button"/>
-	<widget name="Separator"/>
-	<widget name="Space"/>
-	<widget name="Caption"/>
-	<widget name="Checkbox"/>
-	<widget name="Groupbox"/>
-	<widget name="Html"/>
-	<widget name="Popup"/>
-	<widget name="Radio"/>
-	<widget name="Radiogroup"/>
-	<widget name="Toolbar"/>
-	<widget name="Toolbarbutton"/>
-	<widget name="Nodom"/>
-	<widget name="Image"/>
-	<widget name="Imagemap"/>
-	<widget name="Area"/>
-	<widget name="Chart"/>
-	<widget name="Captcha"/>
-	<widget name="Progressmeter"/>
-	<widget name="Fileupload"/>
-	<widget name="Combobutton"/>
-	<widget name="Selectbox"/>
-	<widget name="Notification"/>
-	<widget name="Rating" />
-	<widget name="Inputgroup" />
+	<script src="index.js" jsdoc="true"/>
+
+	<widget name="A" moldOnly="true"/>
+	<widget name="Cell" moldOnly="true"/>	
+	<widget name="Div" moldOnly="true"/>
+	<widget name="Span" moldOnly="true"/>
+	<widget name="Idspace" moldOnly="true"/>
+	<widget name="Include" moldOnly="true"/>
+	<widget name="Label" moldOnly="true"/>
+	<widget name="Button" moldOnly="true"/>
+	<widget name="Separator" moldOnly="true"/>
+	<widget name="Space" moldOnly="true"/>
+	<widget name="Caption" moldOnly="true"/>
+	<widget name="Checkbox" moldOnly="true"/>
+	<widget name="Groupbox" moldOnly="true"/>
+	<widget name="Html" moldOnly="true"/>
+	<widget name="Popup" moldOnly="true"/>
+	<widget name="Radio" moldOnly="true"/>
+	<widget name="Radiogroup" moldOnly="true"/>
+	<widget name="Toolbar" moldOnly="true"/>
+	<widget name="Toolbarbutton" moldOnly="true"/>
+	<widget name="Nodom" moldOnly="true"/>
+	<widget name="Image" moldOnly="true"/>
+	<widget name="Imagemap" moldOnly="true"/>
+	<widget name="Area" moldOnly="true"/>
+	<widget name="Chart" moldOnly="true"/>
+	<widget name="Captcha" moldOnly="true"/>
+	<widget name="Progressmeter" moldOnly="true"/>
+	<widget name="Fileupload" moldOnly="true"/>
+	<widget name="Combobutton" moldOnly="true"/>
+	<widget name="Selectbox" moldOnly="true"/>
+	<widget name="Notification" moldOnly="true"/>
+	<widget name="Rating" moldOnly="true"/>
+	<widget name="Inputgroup" moldOnly="true"/>
 </package>
